refactor(cli): extract session runtime commands
